TcpServer

+++ rocket/net/io_thread_group.cc
@@ -29,7 +29,7 @@ void IOThreadGroup::join() {
 }
 
 IOThread* IOThreadGroup::getIOThread() {
-    if(m_index == m_io_thread_groups.size() || m_index == -1) {
+    if(m_index == (int)m_io_thread_groups.size() || m_index == -1) {
         m_index = 0;
     }
     return m_io_thread_groups[m_index++];
